requirements.txt
pandas
gradio
langchain
faiss-cpu
sentence-transformers
transformers
torch
speechrecognition
langchain-community
groq
python-dotenv
numpy
scikit-learn
openai

